+++ diffsky/__init__.py
@@ -1,7 +1,7 @@
-"""
-"""
+""" """
 
 # flake8: noqa
 
 from ._version import __version__
 from .mass_functions.mc_diffmah_tpeak import mc_subhalos
+from .utils import tw_utils
