+++ diffsky/utils/__init__.py
@@ -1,4 +1,7 @@
 """ """
 
-from .crossmatch_utils import crossmatch  # noqa
-from .utility_funcs import _inverse_sigmoid, _sigmoid  # noqa
+# flake8: noqa
+
+from .crossmatch_utils import crossmatch
+from .tw_utils import _tw_interp_kern, _tw_sig_slope, _tw_sigmoid
+from .utility_funcs import _inverse_sigmoid, _sigmoid
